pyproject.toml
[project]
name = "postdoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.75.0",
    "pathspec>=0.12.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "questionary>=2.1.0",
    "rich>=14.0.0",
    "tiktoken>=0.9.0",
]
